Rewrote the Myster file types system so that types can be added/removed.

+++ com/myster/util/ProgressWindow.java
@@ -165,7 +165,7 @@ public class ProgressWindow extends Frame {
 	}
 	
 	//Standard progress suite
-	public void setValue(int value) {
+	public void setValue(long value) {
 		setValue(value, 0);
     }
